feat(member-search): implement member search component with debounce and auto-select
This adds a new MemberSearch component that allows users to search for members with a debounce feature. It includes functionality to auto-select members based on exact matches and integrates with the existing member selection flow.

feat(product-search): enhance product selection with quick add functionality

This update modifies the ProductSearch component to support a quick add mode, allowing products to be added directly with a default quantity of 1. It also resets the search field after adding a product.
